Split compiler and static tests out into individual test runs
Summary:
Currently we run all of the static and compiler tests all together in 3.10. In 3.12 we split the compiler tests out but all of the static tests get a single process.

This splits it out so that we fully parallelize all of these tests on 3.12 and 3.10. The motivation here is to be able to split out the tests particularly under ref leak mode so we get individualized results especially when the tests crash.

`test_cinderjit` is now moved to a serialize test. It disables the JIT which we don't want to happen in a serial run, and it can potentially interfere with tests that want the JIT to be abled for their runs. `test_cinderjit` has its name fixed on the serialization tests.

`cinderx-ci-python-3_10-tests-dev-nosan-refleak` is now passing.

Reviewed By: martindemello

Differential Revision: D69138398

fbshipit-source-id: e5733aed5e7b86281839744a42bc3e9da1706946