Use thunk object for lazy v-table init
Summary:
New in this version: We call `PyObject_GC_Track` on the thunk in `_PyClassLoader_VTableInitThunk_New`. Also use `Py_XSETREF` in `_PyClassLoader_ReinitVtable` (this last part isn't strictly necessary, but seems like good defense in depth).

Currently we use a tuple object for the state we need for lazily initializing v-tables. But we're going to move to a world where all of these things are callable thunks instead, so this does that for the lazy v-table init.

Reviewed By: martindemello

Differential Revision: D69085513

fbshipit-source-id: ba7437e2aca3e988f80e3bf1d300ace3e54fa12a