Remove primitive return function stub
Summary:
This stub is actually currently not doing anything - we never have static functions with an entry point of `_PyFunction_Vectorcall`, it will either be the JIT entry point or `Ci_StaticFunction_Vectorcall`.

We could potentially change this to check for `Ci_StaticFunction_Vectorcall` and have a new entrypoint which will handle the case when the function is never going to be JITed and do something more optimial (not check if the function got JITed on each call). But given that we haven't been using one of those for a while it seems like it's unnecessary, so this just removes it entirely.

Reviewed By: alexmalyshev

Differential Revision: D69085512

fbshipit-source-id: fa5bc3d3517856b52858711913442d7b6572a6e7