Use unique_ptr for the JIT context and JIT list
Summary:
The addition of the `isCinderModule()` function demonstrated how there are
instances where the cinderjit module was not being properly finalized at
shutdown.  We happened to catch it because of changes to the interned string
table in 3.12 resulting in ASAN errors when post-main destructors ran.

Make post-main destructors even more prevalent by having them try to destroy the
entire JIT context if it still exists.  This will likely result in a crash,
which is what we want.  The 3.10 behavior of silently not crashing when
finalize() wasn't called is worse.

Reviewed By: DinoV

Differential Revision: D69248350

fbshipit-source-id: 04b66cc2c78c888643a70aedfa5ba27318360345