Add static method thunk
Summary:
Similar to other changes in this stack this adds a thunk for calling static functions replacing the existing tuple state. This is vectorcall-able unlike the tuple.

While this is the "static method" thunk it can actually be used for anything that needs to drop `self`, and so we'll use it for those purposes elsewhere as when we have an instance value that we're calling we won't drop self.

Reviewed By: martindemello

Differential Revision: D69085511

fbshipit-source-id: 6bedc0107483fb1770bcc09c6a9017fa518154b9