Add class method thunk
Summary:
Updated to add `PyObject_GC_Track(thunk);` to `_PyClassLoader_ClassMethodThunk_New` and use `Py_XSETREF` to replace `tcs_value` with the class method thunk in _PyVTable_setslot_typecheck.

Removes the tuple based state for class methods and uses a specific class method thunk.

This is starting to show how we're getting better separation of concerns. Our handling of co-routines and class-methods is now orthogonal, with the thunk handling the fixup's to `self` and the outer thunk handling the co-routine type checking behavior.

Reviewed By: martindemello

Differential Revision: D69085507

fbshipit-source-id: 3f9b27d9ce16843805c95bccbb8b7a91fcd096b7