Copy generic type definitions and use PyType_FromSpec for array type
Summary:
During our runtime tests we create multiple runtimes - spinning them up and tearing them down. In 3.12.7 there's been some changes to the interned string table where the strings are no longer immortalized and get released when the runtime shuts down.

Because of this we ready these types once and then their interned strings are deleted on an additional run causing us to crash. Because the type is already ready'd we will not re-initialize it.

Long term we probably want to move to the multi-phased module initialization, but getting back to our types then becomes relatively difficult, and elsewhere we currently only support single phase init. So this just uses the internal APIs which make these static types continue to work for the time being.

So this at least provides us a quick way to get the Python runtime upgrade unblocked - we use `PyType_FromSpec` for the static array type which is relatively simple. We leverage our existing type copying logic for the generic types which are kind of strange.

Each type get initialized when we load the `_static` module and we publish/clobber the global static type object pointers.

One oddity is that the JIT type system is entirely oriented around using maps of static types for doing `Type` to `PyTypeObject*` conversions, so we need an additional fallback for heap types which aren't present in those maps for the static type.

Reviewed By: alexmalyshev

Differential Revision: D69078502

fbshipit-source-id: 9d1d855e98b3717ce98fb121044b2be69ef47325