Replace InvokeStaticMethod with CallInd
Summary:
This gets rid of the specific `InvokeStaticMethod` HIR instruction in-favor of a more general `CallInd` instruction. `CallInd` lets us call any C function using an address which is determined at runtime, so we can load the entry point and then invoke it.

Part of the reason to go w/ a more generic opcode here is that it's what we need later when we'll be switching to something more like `LOAD_METHOD`/`CALL_METHOD` - we'll just have a function pointer to invoke instead of wanting to load from the type object and invoke a given slot. So this will get usage in multiple spots.

It also pushes a lot more of the state into HIR where it could be reasoned about and hit more optimization passes.

There's maybe one interesting thing here - we need to avoid getting the deopt value kind for these function pointers. At the end of the day these aren't meaningful values in the interpreter so they should always represent some extra state that we're tracking the JIT stack.

Reviewed By: martindemello

Differential Revision: D69085569

fbshipit-source-id: 44d11f704bd3128d0a23bd7990b3afbdbe6c9db8