Fix increase in cost in invoking built-in methods
Summary: We no longer are accounting for `PyMethodDescr_Type` in the things that we can directly invoke. This causes us to treat it as a generic type-checked descriptor which is a lot more expensive (because we need to allocate for the type checking of the unknown descriptor). Instead go through the generic load as we know we can directly invoke the method descriptor via vector call.

Reviewed By: czardoz

Differential Revision: D69479025

fbshipit-source-id: 899f817db6bad083bc31cba41b4204195bb2bdb8