Regenerate cases from 3.12 directory
Summary:
Currently we have a forked version of bytecodes.c in CinderX. There's no need for this, and it makes our upgrades more complicated. If we need to change an opcode we should favor adding an override into cinder-bytecodes.

This updates the regen script to pull from third-party, and brings over `SEND_GEN` which is the one opcode we've seem to have updated in bytecodes.c. It then deletes bytecodes.c from CinderX.

Reviewed By: itamaro

Differential Revision: D69539752

fbshipit-source-id: cefaabf8e392ba0ca1d62ac72d47874169270892