Add option to raise Python error on JIT list parsing errors
Summary:
The current behavior for JIT list parser errors is to disable the JIT entirely
and continue executing.  There's a line that gets logged to stderr, but it's
easy to miss.

This diff adds an option to make it a much louder error, which will fire when
`cinderx` is first imported.

Reviewed By: pilleye

Differential Revision: D69597208

fbshipit-source-id: accdf6530f053aa68efd8afe8b6b1049340eed8f