Fix simplifier not using addReference for tuple gets
Summary:
These two uses of `Type::fromObject()` are blindly applied onto an item from an
arbitrary tuple and hope that it outlives the lifetime of the code object.

If you see all other uses of PyTuple_GET_ITEM() in this file, the result is
always incref'd and attached onto the function's reference set.

Reviewed By: DinoV

Differential Revision: D69863030

fbshipit-source-id: 74441983085c4eea8d7101d741af217c9924602f