Add macro for running Python tests with multiple configurations
Summary:
Trying to get better TestX integration with our Python tests.  This sets up the
scaffolding for running a Python test with:

* Python
* Python + CinderX
* Python + CinderX + Jit (Auto)
* Python + CinderX + Jit (All)

across 3.10 and 3.12.

Many of the tests are automatically being skipped for various reasons, e.g. 3.12
and JIT isn't functional right now.

Needed fixes to test_jitlist.py because the test file is flaky when not all test
functions are run in the exact order they're defined in.

Reviewed By: czardoz

Differential Revision: D69883502

fbshipit-source-id: 751a2e2424c3d3ab47b6ad660d7ba1b53d4ae2cd