Fix issue with traceback being freed during __exit__
Summary:
In 3.12 there's a new `WITH_EXCEPT_START` opcode which is getting the traceback from the exception object and then passing it down via a vectorcall, but only after decref'ing the traceback. But the exception object can have the only reference to the traceback! If the receiver of the traceback doesn't incref it, and does something that causes the traceback to be replaced on the exception object, then they have a dangling reference to a freed traceback object.

When our JITed functions receive the traceback they are indeed not increfing it (unlike the interpreter loop), so we lose the references.

Reviewed By: alexmalyshev

Differential Revision: D70216890

fbshipit-source-id: 7d1c5473338be2ee01c70a7841dffd7b3f7288fa