Init cell vars in frame so that f_locals doesn't crash
Summary:
When CPython tries to reify the locals dictionary for a frame it blindly assumes that all cell vars will exist and be populated.

We do currently load the cell vars, but we don't do anything to actually push them into the frame. This adds a step which will initialize the frame object so it contains the cell vars.

Reviewed By: czardoz

Differential Revision: D70270019

fbshipit-source-id: 613d2d25ca34a81486771f69fddfdac1479e1ea9