Allow importing CinderX on unsupported runtimes
Summary:
The cinderx module tries to import the _cinderx module, if it is available.
_cinderx depends on ABI changes that we've made to the Python runtime.  If it
ever gets successfully loaded against a stock runtime, we'll access invalid
memory and crash.  We've avoided this scenario so far by automatically deploying
_cinderx only into packages with supported runtimes.

For library code that has to work across many different runtimes, we'll need to
have it depend on both modules, but have it not crash when it tries to load the
top-level cinderx module on an unsupported runtime.

Reviewed By: SonicField

Differential Revision: D70324660

fbshipit-source-id: 837c4a5adbe9050da69951dc72d5b5ac6d3ecf12