Fix test_compiler.test_cinder tests for 3.12
Summary:
In 3.12 we have the upstream implementatino of inline comprehensions and super. These are tested in the standard library so we don't need our own tests for them.

Also the test_dis width test needs some tweaks to match the 3.12 version. These get all of the `test_compiler.test_cinder` tests passing.

Reviewed By: martindemello

Differential Revision: D70284062

fbshipit-source-id: 8cf12a7b44a7c55299f665eaff53b9d471907512