Preserve registers before initialzing generator frame
Summary: In 3.12 we now initialize the generator frame upon entry to the function. This requires us to spill the registers for the static entry point where we didn't before.  That initialization also replaces `rbp` with a location in the generator object so that we're properly loading/storing the state from there instead of from the CPU stack. We therefore need to first load the address of arguments before we replace rbp and then preserve it across the frame linking.

Reviewed By: martindemello

Differential Revision: D70405380

fbshipit-source-id: a7ea0dcb82a75902f8a26e44c3e974bbb681225d