Setup _cinderx module to have module state and global instance for state
Summary:
We have issues at shutdown because we have static references that survive the Python runtime. This is super problematic with CPython changing the ref count of interned strings to 2 at interpreter shutdown, and then our GlobalCache shuts down at process exit and tries to decref them.

We should ultimately root all of our static state within the _cinderx module.  This is the first step which switches to using `PyModuleDef_Init` to initialize our module instead of `PyModule_Create` and gives us a `CinderModuleState` to stick things in.

This is somewhat problematic because we have a bunch of standalone components which other components rely upon. Ultimately _cinderx needs to depend on all of those components. For us to then root our state in _cinderx and expose it back down to other components means we need to de-couple things. This adds a new `_cinderx-iface` module which will be de-coupled from the implementation and only contain shared interfaces that are used across the system.

The next diff in the stack shows how we'll use this - it adds a new interface into the module and implements that interface in the other components. _cinderx-lib is responsible for publishing the interface as it has dependencies on everything, but other components in the system can use it via the interface definition.

Reviewed By: alexmalyshev

Differential Revision: D70289373

fbshipit-source-id: 36f20c5ebade5baddf5b9969e6618a166c023886