Add IGlobalCacheManager interface and move shutdown to module finalization
Summary:
This is the first usage of getting things more rooted in our CinderX module. We add a new `IGlobalCacheManager` to the `_cinderx-iface` library which is now used when we need to use the global cache. `_cinderx-lib` will create a concrete `GlobalCacheManager` and populate the CinderXState object with the value.

This lets us shutdown the global cache manager with the runtime, and prevents us from crashing when the global cache manager decref's the string objects after the runtime has shutdown.

We still

Reviewed By: alexmalyshev

Differential Revision: D70334428

fbshipit-source-id: e10be9996bb587c9d49a533e4c5ac37f5e802272