Don't use ast_cache when forcibly compiling from source
Summary:
We have an AST cache which is causing S494820.

The issue here is that we may analyze a static module while processing another static module which imports it to know what types it holds. When that happens we end up caching the AST and when an edit happens we re-use the cached AST in the worker even though the module was never materialized as a Python module.

We do need the AST cache so that we don't re-parse files and see 2 different types whe compiling pyc's ahead of time. But `_compile_static` is only called from `load_compiled_module_from_source` and that's only used by the loader's `source_to_code`.

Differential Revision: D70643909

fbshipit-source-id: 8be154eb517a7e0aedcc8cc005c9d64134d986d8