Fix type checking error
Summary: D70643909 introduced this type checking error. We use this in a couple of spots like `buck2 run fbcode//cinderx:python3.10 --  -m cinderx.compiler --static foo.py` but we can muddle through it with the default parameters and the fact that we're mostly using things for truthiness. But this reads the source code and provides it appropriately.

Reviewed By: alexmalyshev

Differential Revision: D70901117

fbshipit-source-id: d3568e9d6800e1d42f51e4602aebfb152d4a1ecd