Allow None as argument to CheckedDict.get
Summary:
The 2nd argument to CheckedDict.get is defined as accepting an Optional[T] but we're currently expecting a non-optional value. When static Python fills in the default value of None we reject the call.

The issue here is that when we do the `_PyClassLoader_CheckParamType` we are using the `elem` to get the parameter information for the type. But the type check here isn't directly against the `V` type, it's for `Optional[V]`. So we can't use `_PyClassLoader_CheckParamType` as the helper here.

Reviewed By: alexmalyshev

Differential Revision: D71006224

fbshipit-source-id: 4cafb4cb5d84b71bd3a932c5e240d8ed10c86ed7