Explicitly init cinderx in compileall
Summary:
I'm not quite sure what's changed, but this used to have `_static` available w/o explicitly importing and initing cinderx. Now it's not available (maybe D70213609?). But it doesn't hurt to be explicit and init this.

Also there's no need for us to do these conditional imports anymore - this is a CinderX specific script.

Reviewed By: alexmalyshev

Differential Revision: D71078358

fbshipit-source-id: 6e745467670c4710b012e57bb18cdc5bde5bf705