Avoid using code object APIs that allocate in 3.12 when deciding to inline a function.
Summary:
Calling PyCode_GetCode on 3.12 will actually return a copy of the de-optimized opcode. This requires an allocation which we can't do when we're doing multi-threaded compile, and we'll try and get the byte code for function inlining.

Instead of making a copy we should just access the adaptive code directly and use our existing de-instrumentation logic.

Similarly we can't use `PyCode_GetCellvars` because it'll lazy allocate some cached data as well. Instead use the same logic to see if we're a cell/free var that the code object does.

Reviewed By: alexmalyshev

Differential Revision: D71001822

fbshipit-source-id: c483aff55fb10f529c30726c9d0c0fbc79f45830