Enable type lookups during simplify for multi-threaded compiles
Summary:
In 3.12 we can't do PyType_Lookup against a type in multi-threaded compile because it needs to look at the interpreter state.

Instead, we can capture the interpreter and modify our existing `typeLookupSafe` to do the lookup in the cached interpreter state when we're in multi-threaded compile.

We also can no longer assert that we didn't get an error because we don't have a PyThreadState to do so. The "good news" is if an error is set we'll now just crash as we don't have a thread state to set it.

Finally, once we're using these types we may end up trying to watch them for deopt patching. We now just have a set of types which we need to watch and we run through them all after we've finished the multi-threaded compile.

Reviewed By: martindemello

Differential Revision: D70990513

fbshipit-source-id: f37a1ca5a58c5c25d7327c9c32cb4b36ada6928c