Don't use test.support.maybe_get_event_loop_policy() in test_static/common.py
Summary:
Pulling in test.support means needing a full installation, which somehow doesn't
work properly with the -no-cinderx test rules.  That needs to be addressed, but
in the meantime we might as reset the policy back to None here.  Since this is
running in the test case's `setUp()` method nothing else should have established
a custom policy yet.

Reviewed By: martindemello

Differential Revision: D71141914

fbshipit-source-id: 1c7343bd8ab99542712ea794f092a9d85c570f67