Replace sys._clear_type_cache with our function which forwards to it
Summary:
We need to be able to clear our caches in tests when we would otherwise clear caches. We replace sys._clear_type_cache with our own function and forward to the normal function.

This exposes an issue in 3.12 where PyObject_GC_Del accesses the type object, and we can have already freed the generic type for the dictionary. So we set the type to PyDict_Type which we'll know will always be around - the only important thing about the type is it has the same characteristics w.r.t. the pre-header layout.

Reviewed By: alexmalyshev

Differential Revision: D71160801

fbshipit-source-id: d7598c28b8bc99805227933a76c9657150f595cb