Deliver events to classloader after shutdown, make cleanup more reliable
Summary:
Currently we don't deliver dict events to the classloader after shutdown of the runtime, which means we fail to get deallocated events for dictionaries and we continue tracking them. This leads to crashes in the runtime tests.

This has started failing in the runtime tests, I'm not sure why exactly it's happening now as presumably this has been here for a little bit since getting the module state object.

It also seems likely that the clearing could be a little pesimistic - if we had circular trash between the type and the dict it could be the case that the dict outlives the type. That'd result in the weakref not being resolved, so this just hardens that.

Ultimately we'll want to have per-runtime class loader state.

Reviewed By: alexmalyshev

Differential Revision: D71588402

fbshipit-source-id: 5110c4d468fafab25afb68734803cd971af38b9b