Fix failing test_strict test
Summary:
This is failing for 2 reasons - one, we don't get a real `freeze_type` because we import it from cinder where it's not available. And then SETUP_FINALLY is no longer in the generated byte code because it's a pseudo op.

Now that we're freezing the type we get better code gen in the `InlinerEliminationStaticTest.InlinedStoreFieldLoadFieldIsEliminated` because we know the type can't be modified.

Reviewed By: alexmalyshev

Differential Revision: D71484626

fbshipit-source-id: 07906cc029fcaee3f29f7eb76e9b73ee16fee202