Fix a bug with checking container membership for C integers
Summary:
We were doing the same strict typechecks for `in` and `not in` that we were
doing for the mathematical comparison operators. We should ideally be doing
those checks on the container type (e.g. check that we have `int64` in
`list[int64]`) but for now we at least do not raise a false error.

NOTE: This may need to be fixed for other C-backed types; this diff is a quick
fix for the CInt case because it is breaking some instagram-server CI tests.

Reviewed By: czardoz

Differential Revision: D69695614

fbshipit-source-id: 8202d21fd2d15cd7b10c1b0accffe61ff613ffa5