Fix minor coroutine incompatibilities
Summary: This is causing some error messages to show up incorrectly, maybe it has some other impacts too. Another issue with the error messages is our type name. `PyType_FromSpec` wants us to provide the module name but we don't want it - so let's just set it after creating the type so that we don't get a warning.

Reviewed By: alexmalyshev

Differential Revision: D71595035

fbshipit-source-id: 41a2fff1ad74774061ff27844b4b092934e95011