Eagerly import `cinderx.compiler.pycodegen`s dependencies
Summary:
We want to make sure that the static compiler is compiled by the C
compiler before we install the source loader. This may not happen if
lazy imports are enabled when the compiler is imported (and we do not
eagerly import our dependencies). There is code in cinder that attempts
to disable lazy imports before installing the source loader (which imports
the static compiler), but it's possible that the static compiler will be
imported lazily before this code is executed. This is currently happening
for new IG devservers:

1. Lazy imports are enabled.
2. `cinderlib.eager_importer.eager_importer` as part of initializing site.
   This transitively imports `cinderx.compiler.pycodegen` with lazy imports
   enabled, so its dependencies are not imported.
3. We install the static source loader, which finds the lazily imported
   `cinderx.compiler.pycodegen` module in sys.modules.
4. We try to compile something as part of an import using the static compiler.
   The static compiler tries to resolve a lazy import of `cinderx.compiler.future`,
   which hasn't been compiled, so we attempt to invoke the static compiler again.

Reviewed By: brittanyrey, DinoV

Differential Revision: D72009374

fbshipit-source-id: 11d4d1da27221bfefb0b34774fe10f1ea8f4e2bb