Disable test_frame tests that are testing interpreter specific details
Summary:
`test_sneaky_frame_object_teardown` is running a finalizer while popping a frame, but the JIT runs the finalizer before popping the frame.
`test.test_frame.TestCAPI.test_getvar` is expecting to be able to use _testinternalcapi to access a local variable, but it's not populated by the JIT.
`test_sneaky_frame_object` is depending on being able to set a trace function to do things at wacky times, but the JIT doesn't support tracing.

Reviewed By: yuqiw4

Differential Revision: D71842052

fbshipit-source-id: 4a7dc73f5fa108b21fc487aaa3b0c8b00d72ea18