Deopt to EXTENDED_ARG offset, not opcode offset
Summary:
Currently we are using `offset` to set the de-opt index. Unfortunately this has us deopting into the offset after the extended arg. We instead need to calculate where the extended arg is.

It's hard to update this anywhere where we're storing it because we also use this frequently to get to the next opcode, and that becomes more complicated. So when we need to get it for the deopt index now we just do the check to go backwards.

Reviewed By: alexmalyshev

Differential Revision: D71573582

fbshipit-source-id: a5a7ad1eefe1792ecf035f7d7eb886681d753d07