Fix multi-threaded compile on debug builds
Summary:
In debug builds the ref count operations will try and access the PyInterpreterState* for tracking the ref count stats. That's not available to us in multi-threaded compile so we need to increment/decrement manually.

Also fixes a couple of spots where we use the generic size function which calls PyErr_Occurred.

Reviewed By: SonicField

Differential Revision: D72014748

fbshipit-source-id: a4fb13871e26a91f073b50df16b8cf9c43eaf99e