Make it harder to add a new HIR instruction without parser support
Summary:
Generate a function that parses an opcode from the name. Then use this to refactor the parsing function to switch on the opcode name instead of comparing string views.

This effectively makes it more difficult to ignore the parser, because it will not compile without the developer adding another case to the switch.

Reviewed By: alexmalyshev

Differential Revision: D72315431

fbshipit-source-id: a028abe3ef2fb5eea28051e13f4b2d455a54b54f