Add AsyncLazyValue to Cinder 3.12
Summary:
Brings Cinder 3.10's AsyncLazyValue into CinderX. The major changes here are:
1) Use PyType_FromSpec to create the types used and move all of those into some `AsyncLazyValueState`.
2) Drop the majority of the v-table entries, these aren't used by AsyncLazyValue directly, only `source_traceback` is used, we keep the implementation of that.
3) C++-ification of the v-tables to use `Ref` and `BorrowedRef`
4) Replaces NULL with nullptr

Most of the definitions are just living in the 'extern "C"' section of the file and only the stuff that interfaces w/ CinderX has been moved into C++ in the `cinderx` namespace.

Existing test cases from Cinder 3.10 are brought over and are passing, they can run on both 3.10 and 3.12.

This diff adds AsyncLazyValue but nothing but the unittests are using it. It'll get enabled in D72327334.

Reviewed By: SonicField

Differential Revision: D72183711

fbshipit-source-id: 43ff4600c1e1a36b2fd392f52035f29f5d29710e