Add explicit visit_list() method to ASTVisitor
Summary:
Getting rid of the places where lists are passed in as nodes to visit().  This
makes `cinderx.ASTVisitor` behave more like `ast.NodeVisitor`, and it makes pyre
types more sensible.  Otherwise everything going in and coming out has to be
typed as `T | Sequence[T] (T <: AST)`.

Added a runtime check to visit_list() to make sure we don't accidentally pass an
individual node to it.

Reviewed By: SonicField

Differential Revision: D72412292

fbshipit-source-id: b7894d1285cf8dd60612ca30bd52f393e4077b71