Newer option to enable/disable the LIR inliner
Summary:
Trying to minimize the number of hidden global variables in the JIT.  Moving the
variable for this to the global `jit::Config`.  Renaming the option to make it
clear this can enable or disable the optimization.  Nothing uses the option
currently, it's just there as a killswitch / for experimentation.

Reviewed By: SonicField

Differential Revision: D72475911

fbshipit-source-id: 531c746a672f28c96f863f408784c37868ba0fa1