Create a cinderjit function to "uncompile" a Python function
Summary:
Add cinderjit.force_uncompile to remove a function from being tracked by cinderjit. The next call will be through the interpreter.

This is needed for future work, but also in order to properly test the specialization opcodes in a jit-all context.

Also add cinderjit.jit_unsuppress to allow a function that was previously suppressed to now be allowed to compile.

Reviewed By: alexmalyshev

Differential Revision: D72461648

fbshipit-source-id: 709fce0c179aca93ee58fe31dc99efaf57a29a6e