Reset interpreter loop hook if cinderx.init() fails
Summary:
Showing up in test_jitlist when run with 3.12.  The
`test_fail_on_parse_error_startup` intentionally creates a Python process that
fails to initialize CinderX, but it crashes shortly after because it still calls
`Ci_EvalFrame` afterwards which will hit asserts on whether `code_extra_index`
is initialized or not.

Reviewed By: martindemello

Differential Revision: D72665990

fbshipit-source-id: eccd9304f654b687d9cd7cff7f83b0619c80c6b6