Have the clang parser search for the correct libclang.so file.
Summary:
Don't assume the libclang dll is named `libclang.so.<version>`; that is not
guaranteed by LLVM and could break across version upgrades. Instead, search
within buck's runtime files for any file starting `libclang.so` and configure
the clang parser to load that.

Also stops setting `llvm-fb: 17` in the PACKAGE file; we can now just use the
default installed version. (Note that this is still hardcoded to 17 by
`fbcode/PACKAGE`, so we do need to override it locally when testing llvm 19.)

Reviewed By: alexmalyshev

Differential Revision: D72812399

fbshipit-source-id: 93bdd2208d79bfbfa1ca090ae0702b307092994a