Add a script to regenerate the cached borrow files.
Summary:
We added a temporary cache of the generated borrow files for T219748977; it is
worth seeing if we can eventually use cached files by default and only run the
generator when we pull new python sources upstream.

This script simply regenerates the (currently unused) cache files; we will need
to add some way of ensuring it is run whenever the `third-party/python/*`
sources are modified before we can start using the cache.

Reviewed By: alexmalyshev

Differential Revision: D72822557

fbshipit-source-id: deb56ffb863e5d6aa180307dd98499f83af91567