Avoid boxing overhead for if isinstance optimizations
Summary: When we optimize the builtin isinstance function, we emit two paths that both box a primitive boolean to return a Python boolean. If we then use that as the predicate for an if statement, then we end up unboxing the bool immediately and comparing against it. This diff avoids that by making the second comparison use the original unboxed bool.

Reviewed By: alexmalyshev

Differential Revision: D73109663

fbshipit-source-id: c56b6a6d389e7f519f22da59a3d3f9567f003d9f