Make stack walking functions in cinder polyfill library work correctly with 3.12
Summary: Prior to 3.12 we had custom functions in Cinder for walking stacks both to make things work well in the presence of shadow frames and to reconstruct async stacks in a way users expected. In 3.12 the shadow stack is not a thing anymore, but we still need special handling for async frames which can't be done with a trivial walk of frames using standard Python API. So, when available, use our custom async stack walking API in Meta Python 3.12.

Reviewed By: martindemello

Differential Revision: D73526579

fbshipit-source-id: 13a5e61772df484fc1a6178d9e3502c61399909d