Deopt to instructions prefixed EXTENDED_ARGs if present
Summary:
If the oparg for a bytecode instruction doesn't fit in 8 bits, it can be extended 8 bits at a time by prefixing the target instruction with `EXTENDED_ARG` instructions. If we deopt we need to restart the interpeter at the start of any related run of `EXTENDED_ARG`s for the cause instruction. Apparently we never did this, but didn't notice until 3.12 added a test which triggers the error.

I've fixed this by moving all handling of `EXTENDED_ARG`s out of `BytecodeInstructionBlock::Iterator` and into `BytecodeInstruction`, and made the API for working with instruction positions more explicit. Overall this should make it harder to ignore/incorrectly handle `EXTENDED_ARG`s.

This implementation needs to be careful to read the bytecode data to find `EXTENDED_ARG`s lazily as `BytecodeInstruction` instances sometimes have invalid positions (e.g. pointing to before the start of code).

Reviewed By: alexmalyshev

Differential Revision: D73883376

fbshipit-source-id: d38b52506240dca4cbd4dce3fc6086d9344e8235