Categorize failing tests with associated open issue
Summary:
This issue makes up about half of the teams failing tests.
This doesn't fix the issue but it at least makes it clear what the problem is.

Essentially in the `sys.executable`, we don't get cinderx bundled into the runtime bc it's just a regular extension module. So, all of these tests fail.

Reviewed By: yoney

Differential Revision: D74609273

fbshipit-source-id: 38a0ec82b39527eb513b487b8196e31250d76c84