Move CallCFunc function pointers to the LIR generator
Summary:
HIR doesn't actually care about these so they don't need to be here.  It's
awkward having them in HIR because it would lead to a dependency cycle with
generators_rt.cpp, which is why we had this hack with manually declaring the
functions.

Simplify the computation for the CallCFunc names as well.

Reviewed By: mpage

Differential Revision: D74497478

fbshipit-source-id: dcd8afa278ed6fefa4288648d28634898b5a95d0