Refactor Disassembler to be thread-safe and to not expose i386-dis
Summary:
I'd like to make i386-dis an optional dependency of cinderx, possibly even swap
it out with a different disassembler library.  It'll be easier to do that if we
make it an implementation detail of Disassembler, rather than have it bleed out
through the API.

While I was here I spotted that we're using `ThreadCompileSerialize` to make it
thread-safe.  That shouldn't be done outside of the implementation, it should be
internally thread-safe.  Slapping a lock over it.

Reviewed By: SonicField

Differential Revision: D74578618

fbshipit-source-id: 947fd7cb1d32dc7a1f9eb9f1248a1d5977a9a45e