Initialize the code object extra data index earlier
Summary:
The index is used to allocate new `CodeExtra` objects for every code object,
which is done whenever a new code object is allocated or encountered in the
CinderX interpreter loop / frame evaluator.  We want this index to be
initialized before we set the frame evaluator.

Create a few new helpers to make this cleaner.

Reviewed By: jbower-fb

Differential Revision: D74612938

fbshipit-source-id: 9dc4dda36acefe90e73751e0c71c10db21de3c1c