Add API to read JIT lists dynamically
Summary:
CinderX reads JIT lists on startup, which makes it difficult to test this code
path.  Add a function for loading the files at runtime.

Updating the JIT list on its own won't guarantee that the functions on the list
will get compiled.  They might have already been loaded and called which would
reset their vectorcall back to the interpreter.  This function needs to re-check
all functions on the Python heap for compatibility with the JIT list.  This is
expensive, but that's okay because loading a JIT list is meant to be done once
in a program.

read_jit_list() ignores getConfig().jit_list.error_on_parse and always raises an
error when an error occurs.  If users call this function they should expect to
handle any errors explicitly.

Update (and rename) append_jit_list() to behave the same way as read_jit_list().
This is safe as it's only used in our tests.

Reviewed By: mpage

Differential Revision: D74953229

fbshipit-source-id: 26a7a4859a0f66e14346732dae1681294aa789fe