Add experimental CMakeLists.txt
Summary:
Trying to use CMake to build the native extension object.  The goal would be to
get 3.12 builds going first, then we can iterate on getting 3.14 up and running.
We're going to skip 3.13.  And despite CinderX having a lot of 3.10 code, it's
actually for what we call "3.10.cinder", our modified internal runtime.  There's
so many changes in that runtime and it's not reasonable to try to make CinderX
support stock 3.10.

**Neither 3.12 nor 3.14 builds yet**.  For 3.12 we still need to get
`Ci_hook_EvalFrame` replaced with `_PyInterpreterState_SetEvalFrameFunc`, and we
need to integrate more files in Interpreter/ into the build.  3.14 has the same
work plus many other outstanding build errors.

Reviewed By: SonicField

Differential Revision: D75466655

fbshipit-source-id: 71d3c279b8662cee493a5b21d4b2594246e2d661